pyproject.toml
[project]
name = "py-atlas"
version = "0.0.1"
description = "This is a template repository for Python projects that use uv for their dependency management."
authors = [{ name = "Florian Maas", email = "alexei.jobfinder@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.12,<3.13"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pydantic[email]>=2,<3",
    "SQLAlchemy>=2,<3",
    "rrdtool>=0.1.16",
    "pillow>=11.3.0",
    "netflix-spectator-py>=1.1.1",
    "flask>=3.1.2",
    "requests>=2.31.0",
    "dearpygui~=1.11;python_version>='3.12'",
    "dearpygui>=1.11.1",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "dbt-duckdb[md]>=1.9.6",
    "dbt-metricflow[dbt-duckdb]==0.8.2",
    "faker>=37.6.0",
]

[project.urls]
Homepage = "https://alex-jobfinder.github.io/atlas/"
Repository = "https://github.com/alex-jobfinder/atlas"
Documentation = "https://alex-jobfinder.github.io/atlas/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",

    "ty>=0.0.1a16",

    "ruff>=0.11.5",

]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rrd_builder"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests", "tests_python_gui"]
addopts = ["--ignore=third_party"]

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true
extend-exclude = ["third_party"]

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    # Prefer `X | Y` in isinstance over tuple; we support older runtimes
    "UP038",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S108"]

[tool.ruff.format]
preview = true

[tool.black]
line-length = 120
target-version = ["py312"]
